+++ user/src/main/java/pl/gucio/enzo/chronica/user/controller/AccountController.java
@@ -8,18 +8,18 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import pl.gucio.enzo.chronica.user.data.request.CreateUserRequest;
+import pl.gucio.enzo.chronica.user.data.request.CreateOrUpdateUserRequest;
 import pl.gucio.enzo.chronica.user.data.response.CreateUserResponse;
-import pl.gucio.enzo.chronica.user.logic.AccountApi;
+import pl.gucio.enzo.chronica.user.logic.AccountService;
 
 @RestController
 @RequestMapping(path = "/api/account")
 @RequiredArgsConstructor
 public class AccountController {
-    private final AccountApi accountApi;
+    private final AccountService accountService;
 
     @PostMapping("/sign-in")
-    public ResponseEntity<CreateUserResponse> createAccount(@RequestBody CreateUserRequest createUserRequest){
-        return accountApi.createUserResponse(createUserRequest);
+    public ResponseEntity<CreateUserResponse> createAccount(@RequestBody CreateOrUpdateUserRequest createOrUpdateUserRequest){
+        return accountService.create(createOrUpdateUserRequest);
     }
 }
