+++ user/src/main/java/pl/gucio/enzo/chronica/user/data/request/CreateOrUpdateUserRequest.java
@@ -1,10 +1,8 @@
 package pl.gucio.enzo.chronica.user.data.request;
 
-import lombok.Builder;
-import lombok.Getter;
 import pl.gucio.enzo.chronica.user.data.dto.AccountDto;
 import pl.gucio.enzo.chronica.user.data.dto.PersonDto;
 
 
-public record CreateUserRequest(AccountDto accountDto, PersonDto personDto) {
+public record CreateOrUpdateUserRequest(AccountDto accountDto, PersonDto personDto) {
 }
