+++ user/src/main/java/pl/gucio/enzo/chronica/user/data/response/ReadUserResponse.java
@@ -1,0 +1,7 @@
+package pl.gucio.enzo.chronica.user.data.response;
+
+import pl.gucio.enzo.chronica.user.data.dto.AccountDto;
+import pl.gucio.enzo.chronica.user.data.dto.PersonDto;
+
+public record ReadUserResponse(AccountDto accountDto, PersonDto personDto, String message) {
+}
