+++ user/src/main/java/pl/gucio/enzo/chronica/user/data/response/UpdateUserResponse.java
@@ -1,4 +1,4 @@
 package pl.gucio.enzo.chronica.user.data.response;
 
-public record CreateUserResponse(String message) {
+public record UpdateUserResponse(String message) {
 }
