+++ user/src/main/java/pl/gucio/enzo/chronica/user/logic/PersonService.java
@@ -1,4 +1,4 @@
-package pl.gucio.enzo.chronica.user.logic.basic;
+package pl.gucio.enzo.chronica.user.logic;
