--- user/src/main/java/pl/gucio/enzo/chronica/user/logic/basic/AccountService.java
@@ -1,19 +1,0 @@
-package pl.gucio.enzo.chronica.user.logic.basic;
-
-
-
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import pl.gucio.enzo.chronica.user.data.entity.AccountEntity;
-import pl.gucio.enzo.chronica.user.data.repository.AccountRepository;
-
-@Service
-@RequiredArgsConstructor
-public class AccountService {
-    private final AccountRepository accountRepository;
-
-    public void create(AccountEntity account){
-        accountRepository.save(account);
-    }
-}
